+++ src/pl/plajer/buildbattle3/handlers/language/Locale.java
@@ -18,6 +18,9 @@
 
 package pl.plajer.buildbattle3.handlers.language;
 
+import java.util.Arrays;
+import java.util.List;
+
 /**
  * @author Plajer
  * <p>
@@ -25,26 +28,30 @@ package pl.plajer.buildbattle3.handlers.language;
  */
 public enum Locale {
 
-  CHINESE_SIMPLIFIED("简体中文", "zh_Hans", "POEditor contributors (Haoting)"),
-  ENGLISH("English", "en_GB", "Plajer"),
-  FRENCH("Français", "fr_FR", "POEditor contributors"),
-  GERMAN("Deutsch", "de_DE", "Tigerkatze"),
-  HUNGARIAN("Magyar", "hu_HU", "POEditor contributors (montlikadani)"),
-  INDONESIAN("Indonesia", "id_ID", "POEditor contributors"),
-  KOREAN("한국의", "ko_KR", "POEditor contributors"),
-  POLISH("Polski", "pl_PL", "Plajer"),
-  SPANISH("Español", "es_ES", "POEditor contributors"),
-  TURKISH("Türk", "tr_TR", "POEditor contributors"),
-  VIETNAMESE("Việt", "vn_VN", "POEditor contributors (HStreamGamer)");
+  CHINESE_SIMPLIFIED("简体中文", "zh_Hans", "POEditor contributors (Haoting)", Arrays.asList("简体中文", "中文", "chinese", "zh")),
+  ENGLISH("English", "en_GB", "Plajer", Arrays.asList("default", "english", "en")),
+  ESTONIAN("Eesti", "et_EE", "POEditor contributors (kaimokene)", Arrays.asList("estonian", "eesti", "et")),
+  FRENCH("Français", "fr_FR", "POEditor contributors", Arrays.asList("french", "francais", "français", "fr")),
+  GERMAN("Deutsch", "de_DE", "Tigerkatze and POEditor contributors", Arrays.asList("deutsch", "german", "de")),
+  HUNGARIAN("Magyar", "hu_HU", "POEditor contributors (montlikadani)", Arrays.asList("hungarian", "magyar", "hu")),
+  INDONESIAN("Indonesia", "id_ID", "POEditor contributors", Arrays.asList("indonesian", "indonesia", "id")),
+  KOREAN("한국의", "ko_KR", "POEditor contributors", Arrays.asList("korean", "한국의", "kr")),
+  POLISH("Polski", "pl_PL", "Plajer", Arrays.asList("polish", "polski", "pl")),
+  RUSSIAN("Pусский", "ru_RU", "POEditor contributors (Mrake)", Arrays.asList("russian", "russkiy", "pусский", "ru")),
+  SPANISH("Español", "es_ES", "POEditor contributors", Arrays.asList("spanish", "espanol", "español", "es")),
+  TURKISH("Türk", "tr_TR", "POEditor contributors", Arrays.asList("turkish", "turk", "türk", "tr")),
+  VIETNAMESE("Việt", "vn_VN", "POEditor contributors (HStreamGamer)", Arrays.asList("vietnamese", "viet", "việt", "vn"));
 
   String formattedName;
   String prefix;
   String author;
+  List<String> aliases;
 
-  Locale(String formattedName, String prefix, String author) {
+  Locale(String formattedName, String prefix, String author, List<String> aliases) {
     this.prefix = prefix;
     this.formattedName = formattedName;
     this.author = author;
+    this.aliases = aliases;
   }
 
   public String getFormattedName() {
@@ -59,4 +66,7 @@ public enum Locale {
     return prefix;
   }
 
+  public List<String> getAliases() {
+    return aliases;
+  }
 }